remove src/main/webapp/course-data directory
The directory being removed contained all our initial course data.
However, we are now using MongoDB and have an entire portion of the
project dedicating to filling the db with good data from ConU's site.
Since the data collected in the webscraping directory replaces the
data in the src/main/webapp/course-data, I'm removing it.

Been a long time comin'.